Here Comes The Big One
JSON file can now be played, with notes that attempt to be in line with the songs both backend and visually. Gerard Way and their 4 directional based sprites appear based on what input you have in game last (the rest will be later dont bully me T.T) Song plays as well.
Implemented Settings Menu, Pause Menu, and added more functionality to the Main Menu. This commit also contains JSON serialization for saving data to files, in this case saving settings to a folder in the new GameData folder.